Removed old ThemeManager and renamed ThemeManager3 -> ThemeManager
