+++ src/Microsoft.DotNet.Wpf/src/PresentationFramework/PresentationFramework.csproj
@@ -1341,7 +1341,7 @@
     <Compile Include="System\Windows\TemplatePartAttribute.cs" />
     <Compile Include="System\Windows\TemplateVisualStateAttribute.cs" />
     <Compile Include="System\Windows\TextPanelProperties.cs" />
-    <Compile Include="System\Windows\ThemeManager3.cs" />
+    <Compile Include="System\Windows\ThemeManager.cs" />
     <Compile Include="System\Windows\FluentThemeState.cs" />
     <Compile Include="System\Windows\ThemeMode.cs" />
     <Compile Include="System\Windows\ThemeModeConverter.cs" />
