+++ src/Microsoft.DotNet.Wpf/src/PresentationFramework/System/Windows/Appearance/WindowBackdropManager.cs
@@ -33,7 +33,7 @@ internal static class WindowBackdropManager
             return false;
         }
 
-        if(!ThemeManager3.IsFluentThemeEnabled && window.ThemeMode == ThemeMode.None)
+        if(!ThemeManager.IsFluentThemeEnabled && window.ThemeMode == ThemeMode.None)
         {
             return false;
         }
