+++ src/Microsoft.DotNet.Wpf/src/PresentationFramework/System/Windows/Controls/DataGridCheckBoxColumn.cs
@@ -35,7 +35,7 @@ namespace System.Windows.Controls
                 if (_defaultElementStyle == null)
                 {
                     Style baseStyle = null;
-                    if (ThemeManager3.IsFluentThemeEnabled)
+                    if (ThemeManager.IsFluentThemeEnabled)
                     {
                         baseStyle = Application.Current.FindResource(typeof(CheckBox)) as Style;
                     }
@@ -67,7 +67,7 @@ namespace System.Windows.Controls
                 if (_defaultEditingElementStyle == null)
                 {
                     Style baseStyle = null;
-                    if (ThemeManager3.IsFluentThemeEnabled)
+                    if (ThemeManager.IsFluentThemeEnabled)
                     {
                         baseStyle = Application.Current.FindResource(typeof(CheckBox)) as Style;
                     }
